PCB Design Arduino Code and Circuit Diagram
his prototype consists of three 8x8” modules. Each module operates independently of each other. Each module consists of 4 “pixels”. Each pixel is 4 inches square and consists of 21 LEDs, two IR emitting diodes, and two IR photodiodes. The mode of sensing is active infrared.
The IR sensor is a simple voltage divider former by two parallel IR photodiodes which are wired reversed biased. When a photodiode is reversed biased (like an LED connected up backwards) it acts like a variable resistor which changes resistance with respect to the amount of light (mainly infrared light) falling on it. The high resistance of the second static value 1 Megohm (R2) provides a voltage divider with the dividing node rising in voltage with an increase in IR light falling onto the photodiodes(R1).

The IR emitters are 3mm infrared LEDs that I got from mouser.com. In order to experience greater sensing range, the IR emitters have to be very bright. Thus we have to push a lot of